+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Configuration now uses [Space Config](https://silverbullet.md/Space%20Config) instead of only [SETTINGS](https://silverbullet.md/SETTINGS) page
 - Improved results page formatting
 
+### Fixed
+
+- Finding multiple matches in a line
+
 ## [2.1.0] - 2024-07-30
 
 ### Added